app/requirements.txt
asgiref==3.6.0
Django==3.2.6
django-shortuuidfield==0.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2022.7.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
